LPD-53134 Creating Base class to reuse code that browses all the tables in the database to apply a cleanup process
